+++ node/src/components/network/event.rs
@@ -5,7 +5,6 @@ use std::{
 };
 
 use derive_more::From;
-use muxink::backpressured::Ticket;
 use serde::Serialize;
 use static_assertions::const_assert;
 use tracing::Span;
@@ -13,8 +12,8 @@ use tracing::Span;
 use casper_types::PublicKey;
 
 use super::{
-    error::{ConnectionError, MessageReaderError},
-    GossipedAddress, Message, NodeId, Transport,
+    error::{ConnectionError, MessageReceiverError},
+    GossipedAddress, Message, NodeId, Ticket, Transport,
 };
 use crate::{
     effect::{
@@ -57,7 +56,7 @@ where
     /// Incoming connection closed.
     IncomingClosed {
         #[serde(skip_serializing)]
-        result: Result<(), MessageReaderError>,
+        result: Result<(), MessageReceiverError>,
         peer_id: Box<NodeId>,
         peer_addr: SocketAddr,
         #[serde(skip_serializing)]