+++ node/src/effect.rs
@@ -108,7 +108,6 @@ use std::{
 
 use datasize::DataSize;
 use futures::{channel::oneshot, future::BoxFuture, FutureExt};
-use muxink::backpressured::Ticket;
 use once_cell::sync::Lazy;
 use serde::{Serialize, Serializer};
 use smallvec::{smallvec, SmallVec};
@@ -142,7 +141,7 @@ use crate::{
         diagnostics_port::StopAtSpec,
         fetcher::{FetchItem, FetchResult},
         gossiper::GossipItem,
-        network::{blocklist::BlocklistJustification, FromIncoming, NetworkInsights},
+        network::{blocklist::BlocklistJustification, FromIncoming, NetworkInsights, Ticket},
         upgrade_watcher::NextUpgrade,
     },
     contract_runtime::SpeculativeExecutionState,