+++ node/CHANGELOG.md
@@ -18,6 +18,7 @@ All notable changes to this project will be documented in this file.  The format
 
 ### Changed
 * The `state_identifier` parameter of the `query_global_state` JSON-RPC method is now optional. If no `state_identifier` is specified, the highest complete block known to the node will be used to fulfill the request.
+* The underlying network protocol has been changed, now supports multiplexing for better latency and proper backpressuring across nodes.